Create set question
Sets Practice
Description

Let’s say you have two lists A and B. Identify the elements which are common in the two lists A and B and return them in a sorted manner. For example 
In this a list has two list elements list1 and list2  -->  [[5,1,3,4,4,5,6,7],[3,3,5,5, 1 ,7 ,2]]   

A = [5,1,3,4,4,5,6,7]    #  list1

B = [3,3,5,5, 1 ,7 ,2]   # list2



Sample Output:

[1,3,5]


If you observe the sample output here you can see that 

Though value 5 is repeated twice in both lists, in the final output it is present only once.
The values are returned in a sorted manner with values increasing.